[border-agent] auto start/stop border agent
This PR mainly allows starting otbr-agent before wpantund is started.

* updated to latest OpenThread
* fixed conversion warnings

+++ src/commissioner/commissioner_argcargv.cpp
@@ -258,7 +258,7 @@ otbrError ParseArgs(int aArgc, char *aArgv[], CommissionerArgs &aArgs)
         steeringLength = (allowAllJoiners ? 1 : kSteeringDefaultLength);
     }
 
-    aArgs.mSteeringData.Init(steeringLength);
+    aArgs.mSteeringData.Init(static_cast<uint8_t>(steeringLength));
 
     if (!allowAllJoiners)
     {
